% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{cor_matrix}
\usage{
cor_matrix(x)
}
\arguments{
\item{year}{What is the correlation matrix of numeric-only data look like in a specific year?}
}
\description{
This function helps to return a correlation matrix of numeric value of an individual year.
}
\examples{
cor_matrix()
}
\keyword{cor_matrix}
