% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Highest_ast.R
\name{highest_ast}
\alias{highest_ast}
\title{Highest Assist}
\usage{
highest_ast(x)
}
\arguments{
\item{year}{Who is the best point guard in a specific year? Defaults to 2000.}
}
\description{
This function allows you to find out the best passers in nba
}
\examples{
highest_ast()
}
\keyword{assist}
