% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oldest.R
\name{oldest}
\alias{oldest}
\title{oldest}
\usage{
oldest(x)
}
\arguments{
\item{year}{Which year do you want to know the oldest player?Defaults to 2000}
}
\description{
This function helps to find out who is the oldest player in a specific year.
}
\examples{
oldest()
}
\keyword{age}
